requirements.txt
requests>=2.28.0
python-dotenv>=1.0.0
playwright>=1.40.0
markdown>=3.5.0
